refactor(android): extract Sleep pure helpers
